+++ oct-macros/src/derive_impl/decode_enum.rs
@@ -35,7 +35,7 @@ pub fn decode_enum(data: DataEnum, repr: Repr, error: Type) -> TokenStream {
 			match variant.fields {
 				Fields::Unit => quote! { Self::#variant_name },
 
-				Fields::Unnamed(_fields) => quote! { Self::#variant_name (#(#commands, )*) },
+				Fields::Unnamed(_fields) => quote! { Self::#variant_name (#(#commands,)*) },
 
 				Fields::Named(fields) => {
 					let field_names = fields
@@ -43,7 +43,7 @@ pub fn decode_enum(data: DataEnum, repr: Repr, error: Type) -> TokenStream {
 						.into_iter()
 						.map(|field| field.ident.unwrap());
 
-					quote! { Self::#variant_name { #(#field_names: #commands, )* } }
+					quote! { Self::#variant_name { #(#field_names: #commands,)* } }
 				},
 			}
 		});