+++ oct/src/encode/output/test.rs
@@ -9,7 +9,6 @@
 #![cfg(test)]
 
 use oct::encode::Output;
-use oct::error::OutputError;
 
 #[test]
 fn test_encode_output() {
@@ -22,35 +21,40 @@ fn test_encode_output() {
 	assert_eq!(output.remaining(), 0x4);
 
 	assert_eq!(output, [].as_slice());
-	assert_eq!(output.write(&[0x04]), Ok(()));
+	output.write(&[0x04]);
 
 	assert_eq!(output.capacity(),  0x4);
 	assert_eq!(output.position(),  0x1);
 	assert_eq!(output.remaining(), 0x3);
 
 	assert_eq!(output, [0x04].as_slice());
-	assert_eq!(output.write(&[0x03]), Ok(()));
+	output.write(&[0x03]);
 
 	assert_eq!(output.capacity(),  0x4);
 	assert_eq!(output.position(),  0x2);
 	assert_eq!(output.remaining(), 0x2);
 
 	assert_eq!(output, [0x04, 0x03].as_slice());
-	assert_eq!(output.write(&[0x02]), Ok(()));
+	output.write(&[0x02]);
 
 	assert_eq!(output.capacity(),  0x4);
 	assert_eq!(output.position(),  0x3);
 	assert_eq!(output.remaining(), 0x1);
 
 	assert_eq!(output, [0x04, 0x03, 0x02].as_slice());
-	assert_eq!(output.write(&[0x01]), Ok(()));
+	output.write(&[0x01]);
 
 	assert_eq!(output.capacity(),  0x4);
 	assert_eq!(output.position(),  0x4);
 	assert_eq!(output.remaining(), 0x0);
 
 	assert_eq!(output, [0x04, 0x03, 0x02, 0x01].as_slice());
-	assert_eq!(output.write(&[0x00]), Err(OutputError { capacity: 0x4, position: 0x4, count: 0x1 }));
+}
 
-	assert_eq!(output.write(&[]), Ok(()));
+#[should_panic]
+#[test]
+fn test_encode_output_empty() {
+	let mut output = Output::new(&mut []);
+
+	output.write(&[0x00]);
 }