+++ oct/src/error/length_error.rs
@@ -27,7 +27,7 @@ pub struct LengthError {
 impl Display for LengthError {
 	#[inline]
 	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-		write!(f, "collection with ({}) remaining size cannot hold ({}) more elements", self.remaining, self.count)
+		write!(f, "collection with `{}` remaining size cannot hold `{}` more elements", self.remaining, self.count)
 	}
 }